+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Override image size for labels and overviews if it is likely to be wrong.
+
 ### Changed
 
 - Use logger instead of issuing warnings.